Update print_int_array() function

+++ src/c/exercist/cyclic_rotation.c
@@ -64,9 +64,9 @@ struct Results {
     int N;
 };
 
-void print_int_array(int arr[], size_t size) {
+void print_int_array(int arr[], int size) {
     printf("[");
-    for (unsigned long i = 0; i < size; ++i) {
+    for (int i = 0; i < size; ++i) {
         if (i == size - 1) printf("%i", arr[i]);
         else printf("%i, ", arr[i]);
     }
