+++ src/c/exercist/tape_equilibrium.c
@@ -52,9 +52,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void print_int_array(int arr[], size_t size) {
+void print_int_array(int arr[], int size) {
     printf("[");
-    for (unsigned long i = 0; i < size; ++i) {
+    for (int i = 0; i < size; ++i) {
         if (i == size - 1) printf("%i", arr[i]);
         else printf("%i, ", arr[i]);
     }
